scripts.js calculator code removed
